perf(ci): Parallelize test execution
This change modifies the CI workflow to run the tests for each Go package in parallel. This is achieved by running each `go test` command in the background and then waiting for all of them to complete. This will significantly reduce the overall test time and speed up the CI process.